Add code and tests to attach adhoc perimeters
