+++ src/geofr/tests/test_admin.py
@@ -1,0 +1,9 @@
+import pytest
+from django.urls import reverse
+
+
+pytestmark = pytest.mark.django_db
+
+
+def test_perimeter_upload(admin_client, perimeters):
+    pass
